Bug is fixed. Im a GODDAMN GENIUS
Bug was from two things:
first: setting initial values from phasespace did not check in which direction to rotate. Now it does using a intersections test to check wether inside. If it fails, it rotates the other direction
second: in gdscript DO NOT USE 0 IN PHASESPACE COORDS. CODE CANNOT HANDLE